+++ website/styles/style.css
@@ -1139,118 +1139,6 @@ input {
     height: 26px;
   }
 
-#bug-report-form {
-    width: 75vw;; /* For phone compatability */
-    max-width: 600px; /* For desktop compatability */
-    z-index: -999;
-    position: fixed;
-    top: 30%;
-    left: 50%;
-    transform: translate(-42%, -50%);
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: 2.4vh;
-    opacity: 0%;
-    transition: all 0.35s ease;
-}
-
-#bug-report-form.active {
-    z-index: 9999;
-    top: 47.5%;
-    opacity: 100%;
-}
-
-#bug-report-form p {
-    font-size: 1.8rem;
-    margin-bottom: 30px;
-}
-
-#bug-report-form label {
-    font-size: 1.2rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start; 
-    text-align: left;
-}
-
-#bug-report-form label input,
-#issue-select,
-#issue-description-field {
-    background-color: #76767639;
-    color: #eee;
-    width: 100%;
-    margin-top: 10px;
-    font-size: 1.1rem;
-    height: 40px;
-    box-shadow: none;
-    border: 2px solid rgba(255, 255, 255, 0.19)
-}
-
-#issue-select {
-    border-radius: 5px;
-}
-
-#issue-select option {
-    color: black;
-    background-color: rgba(255, 255, 255, 0.3);
-}
-
-#issue-description-field {
-    padding: 3px 4px;
-    height: 80px;
-    border-radius: 5px;
-}
-
-#bug-report-form .button-div {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-}
-
-#bug-report-form button[type="submit"] {
-    background-color: rgba(0, 123, 121, 0.364);
-    color: rgb(224, 224, 224);
-    border: none;
-    border-radius: 3px;
-    width: 70px;
-    height: 37px;
-    transition: all 0.2s ease-out;
-}
-
-#bug-report-form button[type="submit"]:hover {
-    color: black;
-    background-color:rgba(0, 240, 236, 0.867);
-    box-shadow: 0px 0px 15px rgba(0, 240, 236, 0.867);
-}
-
-#close-report-popup {
-    position: absolute;
-    background-color: rgba(0, 0, 0, 0);
-    color: rgb(148, 51, 51);
-    border: none;
-    transform: scale(1.3);
-    right: 1%;
-    top: 1%;
-    will-change: transform;
-    transition: transform 0.2s ease;
-}
-
-#close-report-popup:hover {
-    transform: scale(1.4);
-}
-
-body.popup-active {
-    overflow: hidden;
-}
-
-#container.popup-active{
-    user-select: none;
-    pointer-events: none;
-    filter: blur(8px);
-    overflow: hidden;
-  }
-
 /*===========================RESPONSIVE===================================*/
 @media screen and (max-width: 660px) {
     .name {
