+++ website/styles/support.css
@@ -1,0 +1,254 @@
+body.light-mode {
+    --heading-color: rgb(43, 43, 43);
+    --card-box-shadow: 0px 0px 12px rgba(0, 0, 0, 0.7);
+    --card-background-color: rgb(188, 196, 196);
+    --card-font-color: rgba(36, 36, 36, 0.95);
+    --card-p-color: rgb(41, 40, 40);
+    --form-field-input-color: #0f0f0f;
+}
+
+body.dark-mode {
+    --heading-color: rgb(170, 170, 170);
+    --card-box-shadow: 0px 0px 7px rgba(255, 255, 255, 0.7);
+    --card-background-color: rgb(126, 126, 126);
+    --card-font-color: rgba(0, 0, 0, 0.95);
+    --card-p-color: rgb(22, 22, 22);
+    --form-field-input-color: #eee;
+}
+
+main {
+    /* border: 2px solid rgba(0, 0, 0, 0.3); */
+    width: 100%;
+    height: 100vh;
+    position: relative;
+    top: 100px;
+    padding: 5px 10px;
+    display: flex;
+    flex-direction: column;
+    gap: 15vh;
+}
+
+#heading {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 1.45rem;
+    color: var(--heading-color);
+}
+
+section#help-categories {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 5vh;
+}
+
+.category-row {
+    display: flex;
+    gap: 10vw;
+}
+
+.category-row div {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 300px;
+    height: 200px;
+    padding: 15px 10px;
+    background-color: var(--card-background-color);
+    border-radius: 7.5px;
+    box-shadow: var(--card-box-shadow);
+    transition: all 0.3s ease;
+    will-change: transform;
+    cursor: pointer;
+    color: var(--card-font-color);
+    /* border: 2px solid black; */
+}
+
+.category-row div:hover {
+    transform: translateY(-8px);
+}
+
+.category-row div h2 {
+    font-size: 1.95rem;
+    font-weight: 500;
+    
+}
+
+.category-row div p {
+    width: 100%;
+    text-align: center;
+    font-size: 1.5rem;
+    color: var(--card-p-color);
+}
+
+#bug-report-form {
+    width: 75vw;; /* For phone compatability */
+    max-width: 600px; /* For desktop compatability */
+    z-index: -999;
+    position: fixed;
+    top: 30%;
+    left: 50%;
+    transform: translate(-40%, -50%);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    gap: 2.4vh;
+    opacity: 0%;
+    transition: all 0.35s ease;
+}
+
+#bug-report-form.active {
+    z-index: 9999;
+    top: 45%;
+    opacity: 100%;
+}
+
+#bug-report-form p {
+    font-size: 1.8rem;
+    margin-bottom: 30px;
+}
+
+#bug-report-form label {
+    font-size: 1.2rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start; 
+    text-align: left;
+}
+
+#bug-report-form label input,
+#issue-select,
+#issue-description-field {
+    background-color: #76767639;
+    color: var(--form-field-input-color);
+    width: 100%;
+    margin-top: 10px;
+    font-size: 1.1rem;
+    height: 40px;
+    box-shadow: none;
+    border: 2px solid rgba(255, 255, 255, 0.19)
+}
+
+#issue-select {
+    border-radius: 5px;
+}
+
+#issue-select option {
+    color: black;
+    background-color: rgba(255, 255, 255, 0.3);
+}
+
+#issue-description-field {
+    padding: 3px 4px;
+    height: 80px;
+    border-radius: 5px;
+}
+
+#bug-report-form .button-div {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+}
+
+#bug-report-form button[type="submit"] {
+    background-color: rgba(0, 123, 121, 0.364);
+    color: rgb(224, 224, 224);
+    border: none;
+    border-radius: 3px;
+    width: 70px;
+    height: 37px;
+    transition: all 0.2s ease-out;
+}
+
+#bug-report-form button[type="submit"]:hover {
+    color: black;
+    background-color:rgba(0, 240, 236, 0.867);
+    box-shadow: 0px 0px 15px rgba(0, 240, 236, 0.867);
+}
+
+#close-report-popup {
+    position: absolute;
+    background-color: rgba(0, 0, 0, 0);
+    color: rgb(148, 51, 51);
+    border: none;
+    transform: scale(1.3);
+    right: 1%;
+    top: 1%;
+    will-change: transform;
+    transition: transform 0.2s ease;
+}
+
+#close-report-popup:hover {
+    transform: scale(1.4);
+}
+
+body.popup-active {
+    overflow: hidden;
+}
+
+#container.popup-active{
+    user-select: none;
+    pointer-events: none;
+    filter: blur(8px);
+    overflow: hidden;
+  }
+
+/*==============================RESPONSIVE===================================*/
+@media screen and (max-width: 1150px) {
+    .category-row {
+        display: flex;
+        flex-direction: column;
+        gap: 8vw
+    }
+
+    section#help-categories {
+        gap: 8vw;
+    }
+}
+
+@media screen and (max-width: 500px) {
+    #bug-report-form {
+        transform: translate(-50%, -50%);
+    }
+}
+
+@media screen and (max-width: 453px) {
+    #heading {
+        font-size: 1.4rem;
+    }
+
+    #popup-header p {
+        font-size: 1.5rem;
+    }
+}
+
+@media screen and (max-width: 438px) {
+    #heading {
+        font-size: 1.3rem;
+    }
+}
+
+@media screen and (max-width: 408px) {
+    #heading {
+        font-size: 1.2rem;
+    }
+
+    #popup-header p {
+        font-size: 1.4rem;
+    }
+}
+
+@media screen and (max-width: 377px) {
+    #heading {
+        font-size: 1.1rem;
+    }
+}
+
+@media screen and (max-width: 340px) {
+    #popup-header p {
+        font-size: 1.3rem;
+    }
+}
